refactor(plugins): split the install handler into testable phase modules
packages/host/src/api/plugins/install.ts (793 lines, ~500-line post())
becomes a ~100-line sequencing handler over six phase modules under
api/plugins/install/: body validation, the dev-install/link branch,
source resolution (local containment + github parsing + Whiskit artifact
preference), staged-manifest validation, the C13 consent-token gate, and
the commit phase (copy + compile + lockfile record + activation).

Phase bodies move byte-for-byte; every early-return Response, audit
event, and staging-dir teardown keeps its exact semantics. The three
security-critical pieces the audit flagged as buried and untestable now
live in small directly-unit-testable functions with focused tests:

- evaluateConsentGate (consent-token manifestSha binding — valid token
  proceeds, drifted manifest bounces to awaitingConsent+manifestChanged,
  forged/missing/wrong-source tokens 400)
- coreIdSquattingError (core-plugin-id collision rejected without the
  explicit overrideCore opt-in)
- buildSourceInstall (FW1.7 — shipped dist/ deleted BEFORE the builder
  runs so a source install can never execute a prebuilt bundle)
